move utf8 buf to common
